[긴급] 강의 추가시 소숫점 학점 반영

+++ src/main/java/org/wooriverygood/api/course/dto/CourseResponse.java
@@ -8,14 +8,14 @@ import org.wooriverygood.api.course.domain.Courses;
 public class CourseResponse {
     private final Long course_id;
     private final String course_category;
-    private final int course_credit;
+    private final double course_credit;
     private final String course_name;
     private final int isYouguan;
     private final String kaikeYuanxi;
     private final int reviewCount;
 
     @Builder
-    public CourseResponse(Long course_id, String course_category, int course_credit, String course_name, int isYouguan, String kaikeYuanxi, int reviewCount) {
+    public CourseResponse(Long course_id, String course_category, double course_credit, String course_name, int isYouguan, String kaikeYuanxi, int reviewCount) {
         this.course_id = course_id;
         this.course_category = course_category;
         this.course_credit = course_credit;
